move create testcase function to test_util
